Run remove_toggle_ref in main context loop
